requirements.txt
aiofiles==24.1.0
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
appscript==1.1.2
async-generator==1.10
async-timeout==5.0.1
attrs==20.3.0
bandit==1.7.7
bcrypt==4.3.0
beautifulsoup4==4.11.1
billiard==4.2.1
black==24.1.1
blinker==1.9.0
boolean.py==5.0
boto3==1.34.51
botocore==1.34.162
CacheControl==0.14.3
cachetools==5.5.2
celery==5.3.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
check50==3.3.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.8
coverage==7.4.3
cryptography==38.0.4
cyclonedx-python-lib==9.1.0
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
exceptiongroup==1.2.2
fastapi==0.104.1
filelock==3.18.0
flake8==7.0.0
flask-cors==5.0.1
Flask==3.1.0
fpdf==1.7.2
google-api-core==2.24.2
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth==2.39.0
googleapis-common-protos==1.70.0
gotrue==2.9.1
greenlet==3.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
hyperframe==6.1.0
idna==3.3
ifaddr==0.1.7
imageio-ffmpeg==0.6.0
imageio==2.37.0
importlib_metadata==8.6.1
imutils==0.5.4
iniconfig==2.1.0
instagrapi==2.0.1
isort==5.13.2
itsdangerous==2.2.0
jellyfish==0.9.0
Jinja2==3.1.2
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.1
keyboard==0.13.5
kombu==5.5.3
lib50==3.0.4
license-expression==30.4.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mccabe==0.7.0
mdurl==0.1.2
moviepy==1.0.3
msgpack==1.1.0
mypy-extensions==1.0.0
mypy==1.8.0
numpy==2.2.6
obs-websocket-py==0.5.3
obswebsocket-trio==0.1.1.post2
openai==1.3.7
orjson==3.10.16
outcome==1.1.0
packageurl-python==0.16.0
packaging==24.2
pandas==2.2.3
params==0.9.0
passlib[bcrypt]==1.7.4
pathspec==0.12.1
pbr==6.1.1
pexpect==4.8.0
pillow==11.2.1
pip_audit==2.9.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
platformdirs==4.3.7
pluggy==1.5.0
postgrest==0.15.1
proglog==0.1.11
prometheus-client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py-serializable==2.0.0
pyasn1_modules==0.4.2
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.21
pycryptodomex==3.18.0
pydantic_core==2.14.5
pydantic-extra-types==2.10.3
pydantic-settings==2.1.0
pydantic==2.5.0
pyflakes==3.2.0
Pygments==2.19.1
pyobjc-core==8.1
pyobjc-framework-Accounts==8.1
pyobjc-framework-AddressBook==8.1
pyobjc-framework-AdSupport==8.1
pyobjc-framework-AppleScriptKit==8.1
pyobjc-framework-AppleScriptObjC==8.1
pyobjc-framework-ApplicationServices==8.1
pyobjc-framework-AudioVideoBridging==8.1
pyobjc-framework-AuthenticationServices==8.1
pyobjc-framework-AutomaticAssessmentConfiguration==8.1
pyobjc-framework-Automator==8.1
pyobjc-framework-AVFoundation==8.1
pyobjc-framework-AVKit==8.1
pyobjc-framework-BusinessChat==8.1
pyobjc-framework-CalendarStore==8.1
pyobjc-framework-CFNetwork==8.1
pyobjc-framework-CloudKit==8.1
pyobjc-framework-Cocoa==8.1
pyobjc-framework-Collaboration==8.1
pyobjc-framework-ColorSync==8.1
pyobjc-framework-Contacts==8.1
pyobjc-framework-ContactsUI==8.1
pyobjc-framework-CoreAudio==8.1
pyobjc-framework-CoreAudioKit==8.1
pyobjc-framework-CoreBluetooth==8.1
pyobjc-framework-CoreData==8.1
pyobjc-framework-CoreHaptics==8.1
pyobjc-framework-CoreLocation==8.1
pyobjc-framework-CoreMedia==8.1
pyobjc-framework-CoreMediaIO==8.1
pyobjc-framework-CoreMIDI==8.1
pyobjc-framework-CoreML==8.1
pyobjc-framework-CoreMotion==8.1
pyobjc-framework-CoreServices==8.1
pyobjc-framework-CoreSpotlight==8.1
pyobjc-framework-CoreText==8.1
pyobjc-framework-CoreWLAN==8.1
pyobjc-framework-CryptoTokenKit==8.1
pyobjc-framework-DeviceCheck==8.1
pyobjc-framework-DictionaryServices==8.1
pyobjc-framework-DiscRecording==8.1
pyobjc-framework-DiscRecordingUI==8.1
pyobjc-framework-DiskArbitration==8.1
pyobjc-framework-DVDPlayback==8.1
pyobjc-framework-EventKit==8.1
pyobjc-framework-ExceptionHandling==8.1
pyobjc-framework-ExecutionPolicy==8.1
pyobjc-framework-ExternalAccessory==8.1
pyobjc-framework-FileProvider==8.1
pyobjc-framework-FileProviderUI==8.1
pyobjc-framework-FinderSync==8.1
pyobjc-framework-FSEvents==8.1
pyobjc-framework-GameCenter==8.1
pyobjc-framework-GameController==8.1
pyobjc-framework-GameKit==8.1
pyobjc-framework-GameplayKit==8.1
pyobjc-framework-ImageCaptureCore==8.1
pyobjc-framework-IMServicePlugIn==8.1
pyobjc-framework-InputMethodKit==8.1
pyobjc-framework-InstallerPlugins==8.1
pyobjc-framework-InstantMessage==8.1
pyobjc-framework-Intents==8.1
pyobjc-framework-IOSurface==8.1
pyobjc-framework-iTunesLibrary==8.1
pyobjc-framework-LatentSemanticMapping==8.1
pyobjc-framework-LaunchServices==8.1
pyobjc-framework-libdispatch==8.1
pyobjc-framework-LinkPresentation==8.1
pyobjc-framework-LocalAuthentication==8.1
pyobjc-framework-MapKit==8.1
pyobjc-framework-MediaAccessibility==8.1
pyobjc-framework-MediaLibrary==8.1
pyobjc-framework-MediaPlayer==8.1
pyobjc-framework-MediaToolbox==8.1
pyobjc-framework-Metal==8.1
pyobjc-framework-MetalKit==8.1
pyobjc-framework-MetalPerformanceShaders==8.1
pyobjc-framework-ModelIO==8.1
pyobjc-framework-MultipeerConnectivity==8.1
pyobjc-framework-NaturalLanguage==8.1
pyobjc-framework-NetFS==8.1
pyobjc-framework-Network==8.1
pyobjc-framework-NetworkExtension==8.1
pyobjc-framework-NotificationCenter==8.1
pyobjc-framework-OpenDirectory==8.1
pyobjc-framework-OSAKit==8.1
pyobjc-framework-OSLog==8.1
pyobjc-framework-PencilKit==8.1
pyobjc-framework-Photos==8.1
pyobjc-framework-PhotosUI==8.1
pyobjc-framework-PreferencePanes==8.1
pyobjc-framework-PushKit==8.1
pyobjc-framework-Quartz==8.1
pyobjc-framework-QuickLookThumbnailing==8.1
pyobjc-framework-SafariServices==8.1
pyobjc-framework-SceneKit==8.1
pyobjc-framework-ScreenSaver==8.1
pyobjc-framework-ScriptingBridge==8.1
pyobjc-framework-SearchKit==8.1
pyobjc-framework-Security==8.1
pyobjc-framework-SecurityFoundation==8.1
pyobjc-framework-SecurityInterface==8.1
pyobjc-framework-ServiceManagement==8.1
pyobjc-framework-Social==8.1
pyobjc-framework-SoundAnalysis==8.1
pyobjc-framework-Speech==8.1
pyobjc-framework-SpriteKit==8.1
pyobjc-framework-StoreKit==8.1
pyobjc-framework-SyncServices==8.1
pyobjc-framework-SystemConfiguration==8.1
pyobjc-framework-SystemExtensions==8.1
pyobjc-framework-UserNotifications==8.1
pyobjc-framework-VideoSubscriberAccount==8.1
pyobjc-framework-VideoToolbox==8.1
pyobjc-framework-Vision==8.1
pyobjc-framework-WebKit==8.1
pyobjc==8.1
pyparsing==3.2.3
pyserial==3.5
PySocks==1.7.1
pytest-asyncio==1.0.0
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-mock==3.12.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2025.2
PyYAML==5.4.1
realtime==1.0.6
redis==6.2.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.15.100
s3transfer==0.10.4
scikit-learn==1.6.1
scipy==1.13.1
single-source==0.2.0
six==1.16.0
slowapi==0.1.4
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.36.3
stevedore==5.4.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.0.2
supafunc==0.3.3
tenacity
termcolor==1.1.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
trio-websocket==0.9.2
trio==0.19.0
typer==0.4.0
types-cffi==1.17.0.20250326
types-pyasn1==0.6.0.20250208
types-pyOpenSSL==24.1.0.20240722
types-python-jose==3.3.4.20240106
types-redis==4.6.0.20240106
types-requests==2.28.11.17
types-setuptools==78.1.0.20250329
types-urllib3==1.26.25.14
typing_extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.7
uvicorn[standard]==0.24.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.2.1
websockets==12.0
Werkzeug==3.1.3
wiserHeatAPIv2==0.0.7
wsproto==1.0.0
zeroconf==0.37.0
zipp==3.21.0
stripe==7.8.0
sqlalchemy==2.0.23
